Move utils stuff to common package.
